+++ lib/model/fatoora_qr_code_response.dart
@@ -29,4 +29,8 @@ class FatooraQrCodeResponse {
     required this.response,
     required this.status,
   });
+
+  // Map<String, dynamic> toJson(){
+  //   return {"qrCode": qrCode, "status": status, "response" : response.toJson()};
+  // }
 }