+++ CHANGELOG.md
@@ -1,3 +1,4 @@
-## 0.0.1
-
-* First Release.
+## 0.0.2
+Added the file name of the cleared invoice to the Cleared invoice response model.
+Added Cleared Invoice Service for extracting the QR Code and Invoice Hash of the cleared invoice.
+Removed Invoice Hash from the Cleared Invoice response model in favour of Cleared Invoice Service getInvoiceHash Method.